Copied changes over from other repo
I'm a bit rusty with HTML since it's been about a eyar since I've taken the intro to web dev. class, but when playing with the sample CSS and HTML files I've been working on trying to customize them into something more unique

Mainly, I've been working on implementing a frosted glass effect with a more visually appealing background. I'm continuing trying to get it to work as of right now.